Work in progress, training data with loops from DB, currently app still crashes sooner or later because of normalization problem
