+++ packages/backend/src/model/SceneBuilder.js
@@ -1,27 +1,29 @@
 const { trainNetwork } = require("./train");
 const { loadTrack, buildRandomScene } = require("../track");
-const { track } = require("../config");
-// const { buildNewScene, buildRandomScene } = require("./build");
-// const { denormalizeScene } = require("./denormalize");
-// const { normalize } = require("./normalize");
+const { track, sceneQuality, maxAttempts } = require("../config");
+const { Normalizer } = require("./normalize");
 
 module.exports = class {
   constructor({ storage }) {
     this.storage = storage;
     this.trainedNet = null;
     this.track = null;
+    this.normalizer = null;
   }
 
   async init() {
-    const docs = await this.storage.loadLoops();
-    if (docs.length) {
-      this.trainedNet = trainNetwork(docs);
-    }
     this.track = loadTrack(track);
     const {
       meta: { title, copyright }
     } = this.track;
-    console.log(`Track loaded: “${title}”${copyright ? ` – ${copyright}` : ""}`);
+    console.log(
+      `Track loaded: “${title}”${copyright ? ` – ${copyright}` : ""}`
+    );
+    const docs = await this.storage.loadLoops();
+    if (docs.length) {
+      this.normalizer = new Normalizer(docs);
+      this.trainedNet = trainNetwork(this.normalizer.createTrainingData());
+    }
   }
 
   buildRandomScene() {
@@ -32,20 +34,33 @@ module.exports = class {
   }
 
   buildNewScene() {
-    return this.trainedNet
-      ? buildNewScene(this.trainedNet)
-      : this.buildRandomScene();
+    if (!this.track) {
+      throw new Error("You need to call init before building a scene");
+    }
+    if (!this.trainedNet) {
+      return this.buildRandomScene();
+    }
+
+    let scene;
+    let midiCommands;
+    let output;
+    let attempts = 0;
+
+    do {
+      ({ scene, midiCommands } = buildRandomScene(this.track));
+      output = this.trainedNet(this.normalizer.normalizeScene(scene));
+    } while (
+      output.claps - output.boos < sceneQuality &&
+      ++attempts < maxAttempts
+    );
+
+    const { claps, boos } = this.normalizer.denormalizeStats(output);
+
+    console.log(
+      `NEW SCENE PREDICTION:\n${claps.toFixed(0)} claps, ${boos.toFixed(
+        0
+      )} boos (${attempts} attempts)`
+    );
+    return { scene, midiCommands };
   }
-
-  // static buildRandomScene() {
-  //   return buildRandomScene();
-  // }
-
-  // static denormalizeScene(normalizedScene) {
-  //   return denormalizeScene(normalizedScene);
-  // }
-
-  // static normalize(rawData) {
-  //   return normalize(rawData);
-  // }
 };
