--- packages/backend/src/model/denormalize/index.js
@@ -1,3 +1,0 @@
-const denormalizeScene = require("./denormalizeScene");
-
-module.exports = { denormalizeScene };
