--- packages/backend/src/model/denormalize/utils/denormalizeNumber.js
@@ -1,1 +1,0 @@
-module.exports = (value, max) => Math.floor(value * max);
