+++ packages/backend/src/model/normalize/utils/index.js
@@ -1,4 +1,5 @@
 module.exports = {
+  denormalizeValue: require('./denormalizeValue'),
   getNestedProperty: require("./getNestedProperty"),
   normalizeValue: require("./normalizeValue"),
   prettyPrint: require('./prettyPrint')
